requirements-test.txt
# Test dependencies for AquaTru integration
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-homeassistant-custom-component>=0.13.0
pytest-cov>=4.0.0
aiohttp>=3.8.0
